PWGJE: Fix compilation warnings

+++ PWGJE/Tasks/PhotonIsolationQA.cxx
@@ -164,11 +164,12 @@ struct PhotonIsolationQA {
     MC_Info.add("hMcParticleProcessCode", "physical process code of mc particle linked to mc cluster", o2HistType::kTH1F, {{200, 0, 200}});
 
     std::vector<std::string> bin_names = {"A", "B", "C", "D", "True Bckgr A"};
-    for (int i = 0; i < bin_names.size(); i++) {
+    for (size_t i = 0; i < bin_names.size(); i++) {
       MC_Info.get<TH2>(HIST("hABCDControlRegion"))->GetXaxis()->SetBinLabel(i + 1, bin_names[i].c_str());
       Data_Info.get<TH2>(HIST("hABCDControlRegion"))->GetXaxis()->SetBinLabel(i + 1, bin_names[i].c_str());
     }
   }
+
 
   // boolian returns true if a track is matched to the cluster with E_Cluster/P_track < 1.75. Otherwise returns false
   bool track_matching(const auto& cluster, o2::aod::EMCALMatchedTracks const& matched_tracks)
